Refactor object interaction
